Drop compatibility with PHP 5.6, 7.0 and 7.1
