+++ .github/workflows/integrationtest.yml
@@ -26,11 +26,7 @@ jobs:
     strategy:
       matrix:
         include:
-          - php_version: "5.6"
-            wp_version: "6.0"
-            multisite: true
-
-          - php_version: "7.0"
+          - php_version: "7.2"
             wp_version: "latest"
             multisite: false
 
@@ -64,9 +60,9 @@ jobs:
 
     services:
       mysql:
-        # Use MySQL 5.6 for PHP 5.6, use MySQL 5.7 for PHP 7.0 < 7.4, otherwise MySQL 8.0.
+        # Use MySQL 5.7 for PHP 7.2 < 7.4, otherwise MySQL 8.0.
         # Also see: https://core.trac.wordpress.org/ticket/52496
-        image: mysql:${{ ( matrix.php_version == '5.6' && '5.6' ) || ( matrix.php_version < '7.4' && '5.7' ) || '8.0' }}
+        image: mysql:${{ ( matrix.php_version < '7.4' && '5.7' ) || '8.0' }}
         env:
           MYSQL_ALLOW_EMPTY_PASSWORD: false
         ports:
