+++ .github/workflows/lint.yml
@@ -27,7 +27,7 @@ jobs:
       matrix:
         # Lint against the highest/lowest supported versions of each PHP major.
         # And also do a run against "nightly" (the current dev version of PHP).
-        php_version: ["5.6", "7.0", "7.4", "8.0", "8.1", "8.2"]
+        php_version: ["7.2", "7.4", "8.0", "8.1", "8.2"]
 
     name: "Lint: PHP ${{ matrix.php_version }}"
 
@@ -68,16 +68,6 @@ jobs:
           ini-values: zend.assertions=1, error_reporting=-1, display_errors=On
           coverage: none
           tools: cs2pr
-
-      # We don't need to remove and reinstall the Composer dependencies as
-      # the only thing we'll use is the PHP-Parallel-Lint package (PHP 5.3+)
-      # and that means the package is compatible with all supported PHP versions,
-      # so we can just use the version installed on PHP 7.2.
-      # We do need to regenerate the autoload file when switching down as an autoload file
-      # generated by Composer 2.3 (PHP 7.2+) is not usable on PHP < 7.2.
-      - name: Regenerate the Composer autoload file
-        if: ${{ matrix.php_version < '7.1' }}
-        run: composer du --no-scripts
 
       - name: Lint against parse errors
         run: composer lint -- --checkstyle | cs2pr
