+++ .github/workflows/unittest.yml
@@ -25,7 +25,7 @@ jobs:
 
     strategy:
       matrix:
-        php_version: ["5.6", "7.0", "7.2", "7.4", "8.0", "8.1", "8.2"]
+        php_version: ["7.2", "7.4", "8.0", "8.1", "8.2"]
 
     name: "Unit Test: PHP ${{ matrix.php_version }}"
