+++ composer.json
@@ -22,9 +22,8 @@
 		"source": "https://github.com/Yoast/wordpress-seo"
 	},
 	"require": {
-		"php": "^5.6.20 || ^7.0 || ^8.0",
-		"composer/installers": "^1.12 || ^2.0",
-		"yoast/whip": "^1.2"
+		"php": "^7.2 || ^8.0",
+		"composer/installers": "^1.12 || ^2.0"
 	},
 	"require-dev": {
 		"humbug/php-scoper": "^0.13.4",
